+++ sdk/keyvault/azure-security-keyvault-secrets/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## 4.11.0-beta.1 (2026-03-23)
 
+### Features Added
+
+- Added `previousVersion` property to `SecretProperties` which returns the version of the previous secret, if
+  applicable. Applies only to secrets created after June 1, 2025.
+
+### Bugs Fixed
+
+- Fixed an issue where certain `HttpResponseException.getResponse()` calls could cause a `NullPointerException`. ([#47801](https://github.com/Azure/azure-sdk-for-java/issues/47801))
+
 ## 4.10.5 (2026-01-29)
 
 ### Other Changes